refactor: OrderBookRegistrationService 분리 + 오케스트레이터 슬림화 (B-4/B-5)
- C3 OrderBookRegistrationService 신규: registerAfterCommit(afterCommit 등록+isActualTransactionActive 방어 warn-log),
  removeOnCancel(취소 동기 오더북 제거), preSubscribe(시장가 선구독). 무트랜잭션(REQUIRES_NEW 금지 Javadoc 명시).
- OrderService: addOrderToRedisAfterCommit 삭제, RedisOrderBookRepository/OrderSubscriptionCoordinator/
  TransactionHandler 직접 의존 제거 → C3 경유. 메서드레벨 @Transactional 유지, 클래스 Javadoc 갱신(오케스트레이션 역할 명시).
  잔여 주입은 C1~C4 + OrderRepository/AccountRepository/StockRepository만(B-5 목표 이미 충족, 별도 정리 불요).
- 특성화 35/0/0 무수정 green, checkstyle 0, 컨텍스트 기동(SpringBootTest 통합) 확인.
